update to code
added extra functionality that shows the average house prices
